[Chore] 로컬푸쉬와 코드 합치고 정리
