+++ ABloom/ABloom/ABloomApp.swift
@@ -6,21 +6,20 @@
 //
 import Firebase
 import SwiftUI
-import UserNotifications
 
 
 @main
 struct ABloomApp: App {
   // TODO: Notificaiton을 위한 자료
-  //   @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+  @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
   
-  init() {
-    FirebaseApp.configure()
-    // TODO: Local + Firebase 합치기
-    requestNotificationPermission()
-    
-    print("Firebase configured!")
-  }
+  //  init() {
+  ////    FirebaseApp.configure()
+  //    // TODO: Local + Firebase 합치기
+  ////    requestNotificationPermission()
+  //    
+  //    print("Firebase configured!")
+  //  }
   
   var body: some Scene {
     WindowGroup {
